@@ -40,13 +40,13 @@
   "devDependencies": {
     "@bryanrsmith/eslint-config-standard": "^2.1.3",
     "babel-cli": "^6.9.0",
-    "babel-eslint": "^6.1.0",
+    "babel-eslint": "^7.2.2",
     "babel-plugin-istanbul": "^2.0.1",
     "babel-preset-latest": "^6.14.0",
     "babel-preset-stage-2": "^6.13.0",
     "babel-register": "^6.9.0",
     "coveralls": "^2.11.9",
-    "eslint": "^3.0.1",
+    "eslint": "^5.0.1",
     "eslint-config-prettier": "^2.6.0",
     "mocha": "^3.0.2",
     "nyc": "^8.1.0",
